refactor: 리스트 페이지 옵션과 테이블 컴포넌트로 분리
